Added New line and spacing was changed so DBMongo.cpp was also updated.
The real update comes inside ListenerClass.cpp, we now have exact database control from database. I had a prepended string be added when a MongoDB was called which would be really messy in the long term. So now they are designated by the name in the Database. This gives full control to the UI instead of leaving some inside Scribble.